fix par copy
